move to shared: utils.ts
