+++ client/cody-shared/src/utils.ts
