+++ client/cody/src/sourcegraph-api/completions/client.ts
@@ -1,7 +1,7 @@
 import http from 'http'
 import https from 'https'
 
-import { isError } from '../../utils'
+import { isError } from '@sourcegraph/cody-shared/src/utils'
 
 import { parseEvents } from './parse'
 import { Event, CompletionParameters, CompletionCallbacks } from './types'
